chore(core): move `DebugLog` into parent folder
Fixes #7447.
