+++ core/src/meson.build
@@ -62,6 +62,7 @@ lib = library('kmnkbp0',
   'option.cpp',
   'keyboard.cpp',
   'state.cpp',
+  'debuglog.cpp',
   'km_kbp_context_api.cpp',
   'km_kbp_keyboard_api.cpp',
   'km_kbp_options_api.cpp',
@@ -79,7 +80,6 @@ lib = library('kmnkbp0',
   'kmx/kmx_capslock.cpp',
   'kmx/kmx_context.cpp',
   'kmx/kmx_conversion.cpp',
-  'kmx/kmx_debug.cpp',
   'kmx/kmx_debugger.cpp',
   'kmx/kmx_environment.cpp',
   'kmx/kmx_file.cpp',
